scraper/requirements.txt
selenium
psycopg2-binary
beautifulsoup4